+++ src/test/scala/microsites/util/Arbitraries.scala
@@ -131,6 +131,7 @@ trait Arbitraries {
       githubRepo                         ← Arbitrary.arbitrary[String]
       gitHostingService                  ← Arbitrary.arbitrary[GitHostingService]
       gitHostingUrl                      ← Arbitrary.arbitrary[String]
+      gitSidecarChat                     ← Arbitrary.arbitrary[Boolean]
       micrositeKazariEvaluatorUrl        ← Arbitrary.arbitrary[String]
       micrositeKazariEvaluatorToken      ← Arbitrary.arbitrary[String]
       micrositeKazariGithubToken         ← Arbitrary.arbitrary[String]
@@ -161,7 +162,12 @@ trait Arbitraries {
           micrositeExtraMdFiles
         ),
         MicrositeUrlSettings(micrositeBaseUrl, micrositeDocumentationUrl),
-        MicrositeGitSettings(githubOwner, githubRepo, gitHostingService, gitHostingUrl),
+        MicrositeGitSettings(
+          githubOwner,
+          githubRepo,
+          gitHostingService,
+          gitHostingUrl,
+          gitSidecarChat),
         KazariSettings(
           micrositeKazariEvaluatorUrl,
           micrositeKazariEvaluatorToken,